refactor: pull out mda handling logic
